pyproject.toml
[project]
name = "docomo-actualdata-pivot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=3.0.0,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-cloud-bigquery (>=3.40.1,<4.0.0)",
    "google-cloud-storage (>=3.9.0,<4.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "oauth2client (>=4.1.3,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
